Task 3
The vet wants a list of client ids for clients who have one or more
reptiles (defined in the 150A_VetsTables.pdf) .  Display the client id
and the animal type for these clients. The vet does not care how many
animals the clients have, only the client id and types. Sort by the
client id.